# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "ORA Chat API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/chats": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "list chats",
                "responses": {
                    "200": {
                        "description": "A list of chats"
                    }
                }
            },
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "list chats",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name of the Chat",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Authenticated UserId supplied"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/messages/{chat_id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "list of messages",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "formData",
                        "description": "Chat ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of a chat messages"
                    }
                }
            }
        },
        "/api/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "create new message for specific chat",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "formData",
                        "description": "Chat ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "formData",
                        "description": "Message",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Authenticated Chat ID/User ID supplied"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/users": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "update user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "User Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "User Password",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "User Email",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Authenticated UserId supplied"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "create user",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "User Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "User Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "User Email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Authenticated UserId supplied"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/users/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "show user data",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Data"
                    }
                }
            }
        }
    },
    "definitions": {}
}